% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_function.R
\name{assert_function}
\alias{assert_function}
\title{Evaluate conditrion and return error message if confdition not satisfied.}
\usage{
assert_function(condition, msg)
}
\arguments{
\item{condition}{a logical expresssion (e.g. is.null(test_variable))}

\item{msg}{the error message returned if the condition is not met.}
}
\value{
returns tidy dataframe of selected NOMIS table with the selected parameters and user filters applied
}
\description{
Evaluate conditrion and return error message if confdition not satisfied.
}
\examples{
assert_function(1==2,"Incorrect inequality")

}
