% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{availableBoundaries.rda}
\alias{availableBoundaries.rda}
\alias{availableBoundaries}
\title{List of available ONS boundaries}
\format{
A list of with 3348 rows and 1 variable:
\describe{
\item{boundary}{name of boundary table within 'ONS Open geography'}
}
}
\source{
{availableBoundaries.rda}
}
\usage{
availableBoundaries
}
\description{
A list of the available boundary layers that the Office for National Statistics (ONS) has for ArcGIS queries
}
\keyword{data}
