% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sources.R
\name{data_sources}
\alias{data_sources}
\title{Return a list of data sources available on NOMIS}
\usage{
data_sources()
}
\value{
List of all availbale datasources accessible through the Nomis API system. More information can be found here: https://www.nomisweb.co.uk/api/v01/help
}
\description{
Returns a list including the name and id of each
data source available on NOMIS.
}
\examples{
dataSources()

}
