% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dimensions.R
\name{extract_dimensions}
\alias{extract_dimensions}
\title{Extract dimensions available
for a given NOMIS table ID}
\usage{
extract_dimensions(id)
}
\arguments{
\item{id}{A table ID recognised by NOMIS (e.g "NM_1_1")}
}
\value{
A tidy data frame of the dimensions of your chosen nomis table.
}
\description{
Extract dimensions available
for a given NOMIS table ID
}
\examples{
extract_dimensions("NM_1240_1")
}
