% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_layers.R
\name{extract_layers}
\alias{extract_layers}
\title{Retrieve availabe layers for a valid boundary
from the ONS Open Geography Portal}
\usage{
extract_layers(selected_boundary)
}
\arguments{
\item{selected_boundary}{A valid boundary from the ONS Open Geography Portal}
}
\value{
Returns a two dimensional dataframe giving the names and IDs of tables available at the chosen boundary scale.
}
\description{
Retrieve availabe layers for a valid boundary
from the ONS Open Geography Portal
}
\examples{
extract_layers("Middle_Super_Output_Areas_2011_Boundaries")
}
