% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_scales.R
\name{get_available_scales}
\alias{get_available_scales}
\title{Retrieve available spatial scales
for a given NOMIS dataset ID.}
\usage{
get_available_scales(id)
}
\arguments{
\item{id}{A valid NOMIS table id given as a string}
}
\value{
A tidy dataframe listing the Geographical scales available for the NOMIS table selected. E.g. Many census datasets are available at MSOA and LSOA resolutions but not at Regional level
}
\description{
Retrieve available spatial scales
for a given NOMIS dataset ID.
}
\examples{
get_available_scales(id="NM_1003_1")

}
