% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boundaries.R
\name{get_boundaries}
\alias{get_boundaries}
\title{Retrieve boundaries from the ONS Geography Portal
given a valid boundary name and layer name.}
\usage{
get_boundaries(boundary, geometry_filter = NULL)
}
\arguments{
\item{boundary}{A valid boundary name given as a string}

\item{geometry_filter}{geospatial shape or point (using latitude and longitude)
Currently limited to a rectangular box or dropped pin}
}
\value{
Shapefile of all constituencies in the geospatial area submitted through
the geometry_filter, at the chosen ONS Boundary export
}
\description{
Retrieve boundaries from the ONS Geography Portal
given a valid boundary name and layer name.
}
\examples{
get_boundaries("MSOA_Dec_2011_Boundaries_Generalised_Clipped_BGC_EW_V3_2022",
"-1.282825,52.354169,0.206626,52.7106")

}
