% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boundary_postcode.R
\name{get_boundary_postcode}
\alias{get_boundary_postcode}
\title{Function to extract the constituency names for the Postcodes/ latitudes and longitudes of the user's uploaded data}
\usage{
get_boundary_postcode(boundary, layer, column_names, geometry_filter = NULL)
}
\arguments{
\item{boundary}{A valid boundary from the ONS Open Geography Portal}

\item{layer}{ArgGIS layer - defaults to 0}

\item{column_names}{name of field containing constituency names}

\item{geometry_filter}{longitude/latitude coordinates of the point/postcode of interest}
}
\value{
Returns a list of constituency names corresponding to all of the coordinates submitted
}
\description{
Function to extract the constituency names for the Postcodes/ latitudes and longitudes of the user's uploaded data
}
\examples{
get_boundary_postcode("Westminster_Parliamentary_Constituencies_Dec_2022_UK_BGC",0,"PCON22NM","-1.282825,52.354169,0.206626,52.7106")
}
