% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_keyword_table_id.R
\name{get_keyword_table_id}
\alias{get_keyword_table_id}
\title{Function to return list}
\usage{
get_keyword_table_id(usr_keyword)
}
\arguments{
\item{usr_keyword}{Keyword to search nomis tables for}
}
\value{
A tidy dataframe of all nomis tables and their ids, which contain the chosen keyword
}
\description{
Function to return list
}
\examples{
get_keyword_table_id("industry")
}
