% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_overview.R
\name{get_overview}
\alias{get_overview}
\title{Retrieve dataset overview
for a valid NOMIS table id}
\usage{
get_overview(id)
}
\arguments{
\item{id}{A valid NOMIS id.}
}
\value{
json file with overview information of chosen data set - including description of the dataset, last update date, contact for the data
}
\description{
Retrieve dataset overview
for a valid NOMIS table id
}
\examples{
get_overview(id="NM_1_1")
}
