% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table.R
\name{get_table}
\alias{get_table}
\title{Retrieve a NOMIS table using a given table ID.}
\usage{
get_table(id, options, selection = NULL, uid = NULL)
}
\arguments{
\item{id}{a table ID recognised by NOMIS (e.g. "NM_1_1")}

\item{options}{a list of paramaters to pass to the API query.}

\item{selection}{a vector of column names to return. NULL returns all.}

\item{uid}{Unique Nomis identifier to enable larger NOMIS queries - https://www.nomisweb.co.uk/}
}
\value{
returns tidy dataframe of selected NOMIS table with the selected parameters and user filters applied
}
\description{
Retrieve a NOMIS table using a given table ID.
}
\examples{
getTable("NM_1_1", options = list("geography" = "TYPE480", "time" = "latest"))

}
