% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_dimensions.R
\name{get_table_dimensions}
\alias{get_table_dimensions}
\title{Extract dimensions available
for a given NOMIS table ID}
\usage{
get_table_dimensions(id)
}
\arguments{
\item{id}{A table ID recognised by NOMIS (e.g "NM_1_1")}
}
\value{
A tidy data frame of the dimensions of your chosen nomis table.
}
\description{
Extract dimensions available
for a given NOMIS table ID
}
\examples{
get_table_dimensions(id="NM_1240_1")
}
