% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_id.R
\name{get_table_id}
\alias{get_table_id}
\title{Extract unique table id - this is used to
rapidly query census data in other functions}
\usage{
get_table_id(name)
}
\arguments{
\item{name}{A valid NOMIS table name given as a string}
}
\value{
The unique table code that allows rapid querying of the data e.g. "NM_102_1"
}
\description{
Extract unique table id - this is used to
rapidly query census data in other functions
}
\examples{
get_table_id(name="RM003 - Accommodation type by type of central heating in household by tenure")
}
