% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_link_lookup.R
\name{get_table_link_lookup}
\alias{get_table_link_lookup}
\title{Extract table with a lookup between two boundary scales, e.g. get a lookup between Regions and Parliamentary constituencies
for a given NOMIS dataset ID.}
\usage{
get_table_link_lookup(
  lookup_table,
  col_name_1,
  col_name_2,
  col_name_3,
  col_name_4
)
}
\arguments{
\item{lookup_table}{A valid ONS lookup table}

\item{col_name_1}{Field in ONS table containing the constituency code of the smaller scale resolution}

\item{col_name_2}{Field in ONS table containing the constituency code of the larger scale resolution}

\item{col_name_3}{Field in ONS table containing the constituency name of the smaller scale resolution}

\item{col_name_4}{Field in ONS table containing the constituency name of the larger scale resolution}
}
\value{
A dataframe lookup between two chosen boundary resolutions
}
\description{
Extract table with a lookup between two boundary scales, e.g. get a lookup between Regions and Parliamentary constituencies
for a given NOMIS dataset ID.
}
\examples{
get_table_link_lookup("MSOA21_LAD23_EW_LU","MSOA21CD","LAD23CD","MSOA21NM","LAD23NM")

}
