% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_nested.R
\name{is_nested}
\alias{is_nested}
\title{Function to test if list
is nested
#'}
\usage{
is_nested(l)
}
\arguments{
\item{l}{fragment of json file to test}
}
\value{
Boolean indicating whether input was nested or not
}
\description{
Function to test if list
is nested
#'
}
