% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_boundaries.R
\name{list_boundaries}
\alias{list_boundaries}
\title{Retrieve all available ArcGIS boundary layers
from the ONS Open Geography Portal}
\usage{
list_boundaries()
}
\value{
list of avvailable boundary layers on ArcGIS
}
\description{
Retrieve all available ArcGIS boundary layers
from the ONS Open Geography Portal
}
\examples{
list_boundaries()
}
