% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_data_sources.R
\name{list_data_sources}
\alias{list_data_sources}
\title{List 'nomis' Data Sources}
\usage{
list_data_sources(base_url = "https://www.nomisweb.co.uk/api/v01")
}
\arguments{
\item{base_url}{Base nomis url to query}
}
\value{
A tidy dataframe of all available data sources accessible through the 'nomis' API system.
}
\description{
Return a list of the data sources available on 'nomis'.

Returns a list including the name, id and description of each
data source available on 'nomis'. More information can be found here: https://www.nomisweb.co.uk/api/v01/help
}
\examples{
list_data_sources()

}
