% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_tables.R
\name{list_tables}
\alias{list_tables}
\title{List all available datasets on NOMIS}
\usage{
list_tables(base_url = "https://www.nomisweb.co.uk/api/v01/")
}
\value{
data frame containing
the name and ID of each table
}
\description{
List all available datasets on NOMIS
}
\examples{
list_tables()
}
