% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{lookup.rda}
\alias{lookup.rda}
\alias{lookup}
\title{Lookup to match ONS and NOMIS boundary names}
\format{
A data frame with 30 rows and 4 variables:
\describe{
\item{resolution}{boundary layer, written in interpretable manner}
\item{ons}{corresponding boundary layer in the format the ONS will interpret}
\item{nomis}{corresponding boundary layer in the format the ONS will interpret}
\item{Names_and_Codes}{lookup file containing the constituency names and codes of the boundary layer}
}
}
\source{
{lookup.rda}
}
\usage{
lookup
}
\description{
Lookup table providing ONS and NOMIS references for a given ONS boundary layer. The variables are as follows:
}
\keyword{data}
