% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{nomisTables.rda}
\alias{nomisTables.rda}
\alias{nomisTables}
\title{#' List of tables on 'nomis'}
\format{
A data frame with 1605 rows and 3 variables:
\describe{
\item{name}{name of dataset}
\item{id}{code corresponing to dataset}
\item{sourceName}{source of dataset}
}
}
\source{
{nomisTables.rda}
}
\usage{
nomisTables
}
\description{
A dataset containing the available tables, codes and sources of data available on 'nomis'. The variables are as follows:
}
\keyword{data}
