% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_descriptions.R
\docType{data}
\name{scalesForEachDataset.rda}
\alias{scalesForEachDataset.rda}
\alias{scalesForEachDataset}
\title{#' List of tables available at each ONS resolution}
\format{
A data frame with 22564 rows and 3 variables:
\describe{
\item{name}{name of boundary layer}
\item{value}{type of data}
\item{table}{nomis code for the table}
}
}
\source{
{scalesForEachDataset.rda}
}
\usage{
scalesForEachDataset
}
\description{
A dataset containing the available tables on 'nomis', and the boundary layers at which this data is held, e.g. demographic information may only be held in the
Census output areas and not NHS regions. The variables are as follows:
}
\keyword{data}
