% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_id.R
\name{table_id}
\alias{table_id}
\title{Extract unique table id - this is used to
rapidly query census data in other functions}
\usage{
table_id(name, base_url = "https://www.nomisweb.co.uk/api/v01/")
}
\arguments{
\item{name}{A valid NOMIS table name given as a string}
}
\value{
The unique table code that allows rapid querying of the data e.g. "NM_102_1"
}
\description{
Extract unique table id - this is used to
rapidly query census data in other functions
}
\examples{
table_id("RM003 - Accommodation type by type of central heating in household by tenure")
}
