% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_function.R
\name{assert_function}
\alias{assert_function}
\title{Escape Function in Case of Errors}
\usage{
assert_function(condition, msg)
}
\arguments{
\item{condition}{a logical expresssion (e.g. is.null(test_variable))}

\item{msg}{the error message returned if the condition is not met.}
}
\value{
No return value, called for side effects.
}
\description{
Evaluate condition and return error message if condition not satisfied.
}
\examples{
assert_function(1==2,"Incorrect inequality")

}
