% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boundaries_areaname.R
\name{get_boundaries_areaname}
\alias{get_boundaries_areaname}
\title{Extract a geojson file of our chosen constituencies}
\usage{
get_boundaries_areaname(
  boundary,
  layer,
  col_name_var,
  chosen_constituency_list
)
}
\arguments{
\item{boundary}{The resolution of constituencies, e.g. Census Output areas or Westminster Constituencies. Available boundaries can be found here: https://geoportal.statistics.gov.uk/}

\item{layer}{This is equal to 0}

\item{col_name_var}{The name of the datafield where the constituency name is held, e.g. PCON22NM for 2022 Parliamentary Constituencies}

\item{chosen_constituency_list}{List of chosen constituencies}
}
\value{
A geojson spatial file of the constituencies submitted to the function
}
\description{
Extract a geojson file of our chosen constituencies
}
\examples{
get_boundaries_areaname("Local_Authority_Districts_December_2022_UK_BGC_V2")

}
