% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_keyword_table_id.R
\name{get_keyword_table_id}
\alias{get_keyword_table_id}
\title{Keyword Search of 'nomis'}
\usage{
get_keyword_table_id(
  usr_keyword,
  base_url = "https://www.nomisweb.co.uk/api/v01"
)
}
\arguments{
\item{usr_keyword}{Keyword to search 'nomis' tables for, e.g. "religion", "employment", "housing"}

\item{base_url}{nomis url to query}
}
\value{
A tidy dataframe of all nomis tables and their ids, which contain the chosen keyword.
}
\description{
Function to return dataframe of all 'nomis' tables,
and their corresponding codes and descriptions, which have the chosen keyword.
}
\examples{
get_keyword_table_id(usr_keyword="passports")
}
