% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_overview.R
\name{get_overview}
\alias{get_overview}
\title{Table Overview}
\usage{
get_overview(id)
}
\arguments{
\item{id}{A valid 'nomis' id.}
}
\value{
An object with overview information of chosen data set. Object has the structure of the extracted JSON object.
}
\description{
Retrieve dataset overview for a valid 'nomis' table id.
Returned object includes description of the dataset, last update date, contact for the data.
It also extracts all of the available instances of the available table dimensions, which can then be used to filter the dataset 'nomis' table.
}
\examples{
get_overview(id="NM_1_1")

}
