% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_structure.R
\name{get_structure}
\alias{get_structure}
\title{Retrieve dataset overview
for a valid NOMIS table id}
\usage{
get_structure(id, dim)
}
\arguments{
\item{id}{A valid NOMIS id.}

\item{dim}{The name of the g which dimension is queried.}
}
\description{
Retrieve dataset overview
for a valid NOMIS table id
}
\examples{
get_structure("NM_187_1","industry")
}
