% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_dimensions.R
\name{get_table_dimensions}
\alias{get_table_dimensions}
\title{Table Dimensions}
\usage{
get_table_dimensions(id, base_url = "https://www.nomisweb.co.uk/api/v01")
}
\arguments{
\item{id}{A table ID recognised by 'nomis' (e.g "NM_1_1").}

\item{base_url}{Nomis API base url}
}
\value{
A tidy dataframe of the dimensions, and available filtering values, of your chosen 'nomis' table.
}
\description{
Extract dimensions available
for a given 'nomis' table ID. e.g. the dimensions of the table
'RM011 - Country of birth by age ' are age, country and geography;
this function will return all of the available age, country and geography filters available on the table.
}
\examples{
get_table_dimensions(id="NM_1240_1")

}
