% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_id.R
\name{get_table_id}
\alias{get_table_id}
\title{Get 'nomis 'Table IDs}
\usage{
get_table_id(name)
}
\arguments{
\item{name}{A string to search for within 'nomis' table titles.}
}
\value{
A dataframe of 'nomis' table codes and names, as strings, for all 'nomis' tables containing the selected 'name' in their title.
}
\description{
Extract unique table ids for 'nomis' tables containing given name in their title,
these unique table ids can be used to rapidly query census data in other functions e.g. 'get_overview("NM_102_1")'
}
\examples{
get_table_id(name="employment")

}
