% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_info_brief.R
\name{get_table_info_brief}
\alias{get_table_info_brief}
\title{Retrieve summary information about
a given NOMIS dataset.}
\usage{
get_table_info_brief(id)
}
\arguments{
\item{id}{A valid NOMIS table id given as a string, e.g. NM_46_1}
}
\value{
A json file containing the DatasetInfo, DatasetMetadata, Dimensions (variables), Dataset Contact,Units
}
\description{
Retrieve summary information about
a given NOMIS dataset.
}
\examples{
get_table_info_brief("NM_1_1")

}
