% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_boundaries.R
\name{list_boundaries}
\alias{list_boundaries}
\title{List Available Boundaries}
\usage{
list_boundaries(
  base_url = "https://services1.arcgis.com/ESMARspQHYMw9BZ9/arcgis/rest/services"
)
}
\arguments{
\item{base_url}{Open geography portal base url}
}
\value{
A vector of available boundary layers on 'ONS Open Geography'.
}
\description{
Retrieve all available ArcGIS boundary layers
from the 'ONS Open Geography Portal'.
}
